Jobs Summary component
Have the jobs summary be a component
This reduces the already polluted dashboard view
